Made calling audio files more versatile.

+++ SoundBot.js
@@ -1,50 +1,39 @@
 const config = require('./config/default.json');
 const Discord = require("discord.js");
 const client = new Discord.Client();
+var fs = require('fs');
 
-var version = '1.1.2';
+var version = '1.1 DEV';
+var files = fs.readdirSync('./sound');
 client.login(config.token); 
 
 client.on("message", msg => 
 {
 	if (msg.content.startsWith("!"))
 	{
-		var command = msg.content.toLowerCase();
-		switch (command)
+		var command = msg.content;
+		console.log(command);
+		console.log(files.includes(command.substring(1) + '.mp3'));
+		if (files.includes(command.substring(1) + '.mp3'))
 		{
-			case ("!allahu akbar"):
-				playSound(`sound/akbar1.mp3`, msg);
-				break;
-			case ("!terroristswin"):
-				playSound(`sound/csgoTerroristsWin.mp3`, msg);
-				break;
-			case ("!genius"):
-				playSound(`sound/genius.mp3`, msg);
-				break;
-			case ("!whoryou"):
-				playSound(`sound/whoRYou.mp33`, msg);
-				break;
-			case ("!proveit"):
-				playSound(`sound/proveIt.mp3`, msg);
-				break;
-			case ("!ummmno"):
-				playSound(`sound/ummmNo.mp3`, msg);
-				break;
-			case ("!nottoday"):
-				playSound(`sound/notToday.mp3`, msg);
-				break;
-			case ("!now"):
-				playSound(`sound/now.mp3`, msg);
-				break;
-			case ("!helpriley"):
-				listCommands(msg);
-				break;
-			case ("!kys"):
-				client.destroy();
-				break;
-			default:
-				msg.channel.sendMessage("You clearly have no idea what you are doing.  Here is the list of commands.");
-				listCommands(msg);
+			playSound(`sound/` + command.substring(1) + `.mp3`, msg);
+		}
+		else
+		{
+			switch (command)
+			{
+				case ("!HelpRiley"):
+					listCommands(msg);
+					break;
+				case ("!kys"):
+					client.destroy();
+					break;
+				case ("!ListAudio"):
+					msg.channel.sendMessage('Available files: ' + files.toString());
+					break;
+				default:
+					msg.channel.sendMessage('Not a command...');
+			}
 		}
 	}
 });
@@ -52,6 +41,7 @@ client.on("message", msg =>
 client.on('ready', () => 
 {
 	console.log('Rileys Sound v' + version + ' Bot is Running!');
+	console.log('Available files: ' + files.toString());
 });
 
 function playSound(file, msg) 
@@ -83,18 +73,21 @@ function listCommands(msg) {
 	'```',
 	'Riley\'s Sound Bot v' + version,
 	' ',
-	'- Generic Audio -',
+	'- Commands -',
 	'!HelpRiley      --- Show this message',
+	'!ListAudio      --- List all available audio files.',
+	' ',
+	'- Generic Audio -',
 	'!Allahu Akbar   --- Play "ALLAHU AKBAR!" sound',
 	'!TerroristsWin  --- Play CS GO "Terrorists Win" sound',
 	' ',
-	'- Nick\'s Audio -',
+	'- Nick Audio Expansion Pack With Bonus Kody Sound File -',
 	'!Genius         --- Play "Genius!" sound.  Audio courtesy of Nick',
-	'!WhoRYou        --- Play "Who are you?" sound.  Audio courtesy of Nick',
 	'!ProveIt        --- Play "Prove It!" sound.  Audio courtesy of Nick',
 	'!UmmmNo         --- Play "Ummmm no?!?!?" sound.  Audio courtesy of Nick',
 	'!NotToday       --- Play "Not today!" sound.  Audio courtesy of Nick',
 	'!Now            --- Play "Now!" sound.  Audio courtesy of Nick',
+	'!WhoRYou        --- Play "Who are you?" sound.  Audio courtesy of Kody',
 	'```'
 	];
 	msg.channel.sendMessage(message);
